requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
watchdog>=3.0.0
python-barcode>=0.15.0
Pillow>=10.0.0
reportlab>=4.0.0
ttkbootstrap>=1.10.0
pycryptodome>=3.19.0
beautifulsoup4>=4.12.0
requests>=2.31.0
packaging>=23.0
python-docx>=1.0.0
pywin32>=306; sys_platform == 'win32'
pyinstaller>=6.0.0

# Optional — thermal receipt printing
# python-escpos>=3.0a9
